Actualizacion y correccion de errores
Corregimos algunos errore, organizamos metodos y actualizamos
