+++ src/app/interfaces/response.interface.ts
@@ -1,4 +1,4 @@
 export interface ResponseI {
   status: string;
-  response: string;
+  result: string | any;
 }
